requirements.txt
fastapi==0.110.0
uvicorn~=0.29.0
SQLAlchemy~=2.0.29
python-dotenv~=1.0.1
pydantic~=2.6.4